chore: remove custom hook tm-94
