--- apps/frontend/src/pages/friends/hooks/hooks.ts
@@ -1,1 +1,0 @@
-export { useFriendInteractions } from "./use-friend-interactions/use-friend-interactions.js";
